feat: Add video_url parameter to addSoundToList function
This commit modifies the addSoundToList function in the SymbolCard component to include a video_url parameter. If the objectSound has a video_url property, it will be passed along with the user_id and symbol_id when calling the addSoundToList function. This enhancement ensures that the video URL is included when adding a sound to the list.

Note: This commit message follows the convention used in the recent repository commits.